DDF-3649 Extracting our shareability to singular module
