+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/security/ShareableMetacardSharingPolicyPlugin.java
@@ -37,6 +37,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.commons.lang3.StringUtils;
+import org.codice.ddf.catalog.ui.sharing.ShareableMetacardImpl;
 
 public class ShareableMetacardSharingPolicyPlugin implements PolicyPlugin {
