+++ catalog/ui/catalog-ui-search/src/test/java/org/codice/ddf/catalog/ui/security/ShareableMetacardImplTest.java
@@ -24,6 +24,7 @@ import ddf.catalog.data.impl.types.SecurityAttributes;
 import java.util.Collections;
 import java.util.Set;
 import org.codice.ddf.catalog.ui.metacard.workspace.WorkspaceConstants;
+import org.codice.ddf.catalog.ui.sharing.ShareableMetacardImpl;
 import org.junit.Before;
 import org.junit.Test;
