+++ catalog/ui/pom.xml
@@ -33,6 +33,7 @@
         <module>catalog-ui-search</module>
         <module>search-ui-app</module>
         <module>zip-splitter</module>
+        <module>catalog-ui-metacard-sharing-impl</module>
     </modules>
     <properties>
         <cometd.version>3.0.9</cometd.version>
